Updated fetch_google drive and added instuctions Changed .gitignore to not ignore test.txt and trian.txt

+++ src/fetch_google_drive.py
@@ -1,10 +1,10 @@
-#Joshua Grindstaff
-#IMPORTANT need to add service account email to drive, think we'll need to contact client for that
 from google.oauth2 import service_account
 from googleapiclient.discovery import build
 from googleapiclient.errors import HttpError
 from googleapiclient.http import MediaIoBaseDownload
 import os
+from pathlib import Path
+SKIP_EXIST_DIR = 1
 SCOPES = ["https://www.googleapis.com/auth/drive.readonly"]
 #TODO Replace with CreditialFile (Same as test2)
 SERVICE_ACCOUNT_FILE = '.json'
@@ -12,9 +12,9 @@ credentials = service_account.Credentials.from_service_account_file(
     SERVICE_ACCOUNT_FILE, scopes=SCOPES)
 #This is the ID of a test drive I made
 #TODO Change to clients drive
-SUBMISSION_FOLDER_ID = "1o0w7mOrhZh5tCuaVXnpBa8195J5iUBz2"
+SUBMISSION_FOLDER_ID = "1-SFFGspWjvHu0JKsqoWUinpdZtRAqs1v"
 # path to where the test and training data will be created if it does not exists
-DATA_PATH = 'submissions'
+DATA_PATH = '\submissions'
 # How many training images for each test image
 TRAIN_RATIO = 4
 test_path = DATA_PATH + '/test'
@@ -31,12 +31,14 @@ with open(train_txt_path, 'r') as f:
     train_num = int(lines[0])
     #List of assignments (Labeled by dates) that should go into trainning folder
     train  = [line.strip() for line in lines[1:]]
+print(train)
 
 with open(test_txt_path, 'r') as f:
     lines = f.readlines()
     test_num = int(lines[0])
     #List of assignments (Labeled by dates) that should go into trainning folder
     test  = [line.strip() for line in lines[1:]]
+print(test)
 
 def updateTxtFile():
     with open(train_txt_path, 'w') as f:
@@ -58,23 +60,28 @@ def download_folder(folder_id, save_path):
         file_id = file['id']
         file_name = file['name']
         file_mime_type = file.get('mimeType', '')
+        next = Path(f"{current_dir}/{file_name}")
         if file_mime_type == 'application/vnd.google-apps.folder':
-            os.makedirs(current_dir + "/" + file_name, exist_ok=True)  
-            download_folder(file_id, current_dir + "/" + file_name)
+            
+            os.makedirs(next, exist_ok=True)  
+            download_folder(file_id, next)
         else:
-            download_file(file_id, current_dir + "/" + file_name)
+            download_file(file_id, next)
 
     
 #Function to download a specific file to a specific local directory
 def download_file(file_id, save_path):
     try:
-        request = drive_service.files().get_media(fileId=file_id)
-        with open(save_path, 'wb') as f:
-            downloader = MediaIoBaseDownload(f, request)
-            done = False
-            while done is False:
-                status, done = downloader.next_chunk()
-                print(f"Downloading {file_id}: {int(status.progress() * 100)}%")
+        request = drive_service.files().get_media(fileId=file_id)      
+        if not save_path.exists():
+            with open(save_path, 'wb') as f:
+                downloader = MediaIoBaseDownload(f, request)
+                done = False
+                while done is False:
+                    status, done = downloader.next_chunk()
+                    print(f"Downloading {file_id}: {int(status.progress() * 100)}%")
+        else:
+            print(f"{save_path} exists, download skipped")
     except HttpError as error:
         print(f"An error occurred: {error} \nFile {file_id} failed")
         return None
@@ -87,18 +94,25 @@ for file in files:
     file_id = file['id']
     file_name = file['name']
     if file_name in train:
-        download_folder(file_id,DATA_PATH + "/train/" + file_name)
+        file_path = Path(f"{DATA_PATH}/train/{file_name}")
+        if file_path.is_dir() and SKIP_EXIST_DIR:
+            print(f"{file_path} directory already exists... skipping download")
+            continue
     elif file_name in test:
-        download_folder(file_id,DATA_PATH + "/test/"+ file_name)
+        file_path = Path(f"{DATA_PATH}/test/{file_name}")
+        if file_path.is_dir() and SKIP_EXIST_DIR:
+            print(f"{file_path} directory already exists... skipping download")
+            continue
     elif train_num // TRAIN_RATIO >= test_num:
-        os.makedirs(DATA_PATH + "/test/"+ file_name, exist_ok=True)  
-        download_folder(file_id,DATA_PATH + "/test/"+ file_name)
+        file_path = Path(f"{DATA_PATH}/test/{file_name}")
+        os.makedirs(file_path, exist_ok=True)  
         test_num = test_num + 1
         test.append(file_name)
     else:
-        os.makedirs(DATA_PATH + "/train/"+ file_name, exist_ok=True)  
-        download_folder(file_id,DATA_PATH + "/train/"+ file_name)
+        file_path = Path(f"{DATA_PATH}/train/{file_name}")
+        os.makedirs(file_path, exist_ok=True)  
         train_num = train_num + 1   
         train.append(file_name)
+    download_folder(file_id,file_path)
 updateTxtFile()
